+++ src/test/groovy/nebula/plugin/bintray/NebulaOJOPublishingPluginSpec.groovy
@@ -17,9 +17,17 @@ package nebula.plugin.bintray
 
 import nebula.test.PluginProjectSpec
 
-class BintrayPluginSpec extends PluginProjectSpec {
+class NebulaOJOPublishingPluginSpec extends PluginProjectSpec {
     @Override
     String getPluginName() {
-        'nebula-bintray'
+        'nebula-ojo-publishing'
+    }
+
+    def 'apply plugin'() {
+        when:
+        project.plugins.apply('nebula-ojo-publishing')
+
+        then:
+        project.tasks.getByName('artifactoryPublish') != null
     }
 }