+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-1.12.0 / 2014-xx-xx
-==================
+1.12.1 / 2014-07-10
+===================
+
+* Rename to nebula-bintray-plugin
+* Extracted bintray logic from nebula-plugin-plugin
+
+1.12.0 / 2014-06-12
+===================
 
 * Initial release